Store empty values instead of "N/A"
"N/A" shouldn't be stored in the table. It should just store blank values. It used to not work because we switched from empty strings to empty arrays. Now I check if the array has no elements to show "N/A"
I also need to hide the remove button when the user isn't meant to remove claims/evidence & add commas between elements